Don't include NodeUI and Tequilapi in mobile
